Two minor fixed to data types
Fixed a rounding error caused by storing population as an integer (now a double), and an integer overflow with total forage (now a long -- this is probably plenty but it could be unsigned if needed)

+++ src/EventsLog.cs
@@ -25,7 +25,7 @@ namespace Landis.Extension.Browse
         public double PopulationDensity { get; set; }
 
         [DataFieldAttribute(Desc = "Total Forage (kg)")]
-        public int TotalForage { get; set; }
+        public long TotalForage { get; set; }
 
         [DataFieldAttribute(Desc = "Mean Forage (g m^-2)")]
         public int MeanForage { get; set; }
